Add soft-dependency on resize-observer-polyfill
In order to know when a parent element changes its size, listening to
window resize isn't enough. We need to also react to dom resizing
happening as a result of something else. @lencioni pointed me at
https://github.com/que-etc/resize-observer-polyfill and an initial test
is showing good promise. Both the "In a table" and "When styling is
applied asynchronous" examples are now doing the right thing.

I haven't made up my mind yet whether to bundle resize-observer-polyfill
by default or allow it to be injected for those who care. It solves an
important use-case, but everyone is not going to need it. So for now
it's injected when withAvailableWidth is called:

  import ResizeObserver from 'resize-observer-polyfill';
  withAvailableWidth(Foo, { ResizeObserver });