POC: Avoid unmounting to get new width
This commit isn't meant to be merged, as it has a few issues. But I
wanted to push it to a branch anyway, as there are some ideas here that
are worth exploring. For once, it would be good to find a way to reuse
the rendered & mounted component on resize changes.