WIP: Avoid unmount+mount
Instead of throwing out the child on resize events (which causes state
to be lost) we can use the dom element of the child during the update
phase.

The way we do this is quite complicated, and involves a few tricks:

- On mount, remember the index inside the parent container where we drop
the temporary measuring div.
- On first update, remember the dom node of the wrapped component
- On resize events, compute new width on the memoized child

The downsides are:

- Complexity - this is arguably much trickier than an unmount+mount
- The child component needs to expose a DOM element
- If the order of dom element changes between first render and first
update, we might measure the wrong element
- If the child throws out its child for another one, we're holding on to
the wrong element. This is hard to detect.
- If the child sets a static width based on the available width, the
resize events will never cause any changes to `availableWidth`.

The benefit of this approach is that we won't throw out state for the
child.