Use fragments to avoid unmount
On my life-long, rainbow-chasing quest to find a solution to the issue
of having to unmount + mount during a resize, I've come up with a new
solution using React 16's new fragment support [1]. Instead of throwing
out the rendered component, we can hide it using an adjacent sibling
selector (.child-a + *). We drop in the measurement div as usual
on mount. But on resize, we inject some css to temporarily hide the
component, then force a new measurement, then re-render the component
with the new width.

Solves the same problem as #5, but in a more elegant/robust way.

The pros:
- No mounting+unmounting of the child

The cons:
- Leaves the measurement div in the DOM, could be confusing (it's
display: none though).
- Uses adjacent sibling selector, which means that this will work best
with nested components with only one child. I tried using a general
sibling selector (.child-a ~ *) but then I risk affecting anything below
the rendered component.
- Makes the code arguably more complex. There's a counter to force the
measurement div to trigger its ref, there's a generated instance id so
that we can dynamically create a <style> tag locally. Hopefully it's all
worth it in the end.

[1] https://reactjs.org/blog/2017/09/26/react-v16.0.html#new-render-return-types-fragments-and-strings