Don't recompute when only height changes
I'm tracking down a bug on happo.io where a withAvailableWidth container
is loosing its scroll position on recomputing the width. While the
height fix applied in a previous commit will be sufficient, we can also
make sure that the width is only recomputed if the width of the window
changes.